Refactor: Move integration tests to proper location
